+++ src/gdf/graphics/sprite.py
@@ -28,8 +28,8 @@ class Sprite2D:
     def render(self,
                x:      float,
                y:      float,
-               width:  int = 0,
-               height: int = 0,
+               width:  float = 0,
+               height: float = 0,
                angle:  float = 0.0,
                color:  list = None
                ) -> "Sprite2D":
